Array populating!
Created a 'testing' directory with copies to strip down the javascript and start from the beginning with makeWoke and wokeItems. Array is populating now! Going to back-build, adding one-by-one the elements I had working so I can track where, if at all, I'm breaking things and fix them from that level. (Would a branch be best for this? Something to look into later.)